encontre algunos errores muy tontos pero estaria bueno corregirlos

+++ Kernel/interruptions/sysHandler.h
@@ -13,7 +13,7 @@ void sysHandler(uint64_t sysNumber, uint64_t r1, uint64_t r2, uint64_t r3, uint6
 void read(unsigned char * r1, unsigned int r2);
 void getTimeRTC(uint64_t r1, uint64_t r2);
 void getReg(uint64_t * r1, uint64_t * stackFrame);
-void getMem(uint32_t * dir, uint32_t * vec);
+void getMem(uint8_t * dir, uint8_t * vec);
 void clearScreen(int currentScreen);
 void getInfo(uint32_t * r1, uint32_t * r2, int * id);
