+++ Userland/SampleCodeModule/include/commands.h
@@ -28,7 +28,7 @@ int checkCommand(char * buffer, char * parameter);
 extern int _quadratic(long double * a, long double * b, long double * c, long double * r1, long double * r2);
 extern void _exc6Trigger();
 extern void _getReg(uint64_t registers);
-extern void _getMem(uint32_t * dir, uint32_t * memory);
+extern void _getMem(uint8_t * dir, uint8_t * memory);
 extern void _getTime(int * date, int * hours);
 extern void _clearScreen(int cB);
 extern int _getCpuInfo(uint32_t * r1, uint32_t * r2, int * features_id_and_ret);
